pyproject.toml
[project]
name = "02-review-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "openai>=2.33.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "uvicorn>=0.44.0",
]
